Move AbstractInstrumentBuilder to AbstractInstrument.Builder (#884)
