+++ sdk/src/main/java/io/opentelemetry/sdk/metrics/AbstractInstrument.java
@@ -16,7 +16,13 @@
 
 package io.opentelemetry.sdk.metrics;
 
+import io.opentelemetry.internal.StringUtils;
+import io.opentelemetry.internal.Utils;
 import io.opentelemetry.metrics.Instrument;
+import io.opentelemetry.sdk.common.InstrumentationLibraryInfo;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -88,5 +94,92 @@ abstract class AbstractInstrument implements Instrument {
     result = 31 * result + constantLabels.hashCode();
     result = 31 * result + labelKeys.hashCode();
     return result;
+  }
+
+  abstract static class Builder<B extends Instrument.Builder<B, V>, V>
+      implements Instrument.Builder<B, V> {
+    /* VisibleForTesting */ static final int NAME_MAX_LENGTH = 255;
+    /* VisibleForTesting */ static final String ERROR_MESSAGE_INVALID_NAME =
+        "Name should be a ASCII string with a length no greater than "
+            + NAME_MAX_LENGTH
+            + " characters.";
+
+    private final String name;
+    private final MeterSharedState meterSharedState;
+    private final InstrumentationLibraryInfo instrumentationLibraryInfo;
+    private String description = "";
+    private String unit = "1";
+    private List<String> labelKeys = Collections.emptyList();
+    private Map<String, String> constantLabels = Collections.emptyMap();
+
+    Builder(
+        String name,
+        MeterSharedState meterSharedState,
+        InstrumentationLibraryInfo instrumentationLibraryInfo) {
+      Utils.checkNotNull(name, "name");
+      Utils.checkArgument(
+          StringUtils.isValidMetricName(name) && name.length() <= NAME_MAX_LENGTH,
+          ERROR_MESSAGE_INVALID_NAME);
+      this.name = name;
+      this.meterSharedState = meterSharedState;
+      this.instrumentationLibraryInfo = instrumentationLibraryInfo;
+    }
+
+    @Override
+    public final B setDescription(String description) {
+      this.description = Utils.checkNotNull(description, "description");
+      return getThis();
+    }
+
+    @Override
+    public final B setUnit(String unit) {
+      this.unit = Utils.checkNotNull(unit, "unit");
+      return getThis();
+    }
+
+    @Override
+    public final B setLabelKeys(List<String> labelKeys) {
+      Utils.checkListElementNotNull(Utils.checkNotNull(labelKeys, "labelKeys"), "labelKey");
+      this.labelKeys = Collections.unmodifiableList(new ArrayList<>(labelKeys));
+      return getThis();
+    }
+
+    @Override
+    public final B setConstantLabels(Map<String, String> constantLabels) {
+      Utils.checkMapKeysNotNull(
+          Utils.checkNotNull(constantLabels, "constantLabels"), "constantLabel");
+      this.constantLabels = Collections.unmodifiableMap(new HashMap<>(constantLabels));
+      return getThis();
+    }
+
+    final String getName() {
+      return name;
+    }
+
+    final MeterSharedState getMeterSharedState() {
+      return meterSharedState;
+    }
+
+    final InstrumentationLibraryInfo getInstrumentationLibraryInfo() {
+      return instrumentationLibraryInfo;
+    }
+
+    final String getDescription() {
+      return description;
+    }
+
+    final String getUnit() {
+      return unit;
+    }
+
+    final List<String> getLabelKeys() {
+      return labelKeys;
+    }
+
+    final Map<String, String> getConstantLabels() {
+      return constantLabels;
+    }
+
+    abstract B getThis();
   }
 }
